Sales--Email Threshold in processing..

See merge request telliant-mammoth/strive!415

+++ Sourcecode/API/Strive.BusinessEntities/DTO/Product/ProductDto.cs
@@ -19,7 +19,7 @@ namespace Strive.BusinessEntities.ViewModel.Product
         public int? VendorId { get; set; }
         public int? Size { get; set; }
         public string SizeDescription { get; set; }
-        public short? Quantity { get; set; }
+        public decimal? Quantity { get; set; }
         public string QuantityDescription { get; set; }
         public decimal? Cost { get; set; }
         public decimal? Price { get; set; }
